refactor swaps
