+++ src/kong_backend/src/swap/mod.rs
@@ -1,3 +1,6 @@
+pub mod calculate_amounts;
+pub mod return_pay_token;
+pub mod send_receive_token;
 #[allow(clippy::module_inception)]
 pub mod swap;
 pub mod swap_amounts;
@@ -8,3 +11,4 @@ pub mod swap_reply;
 pub mod swap_reply_impl;
 pub mod swap_transfer;
 pub mod swap_transfer_from;
+pub mod update_liquidity_pool;
